Share WorldService evaluation schemas
